src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/HTMLTemplateElement/custom-element-wrapper-gc-expected.txt

CONSOLE ERROR: line 14: Uncaught TypeError: document.querySelector(...).content.ownerDocument.registerElement is not a function
Custom element wrappers in templates should be retained

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL Unexpected error: Uncaught TypeError: document.querySelector(...).content.ownerDocument.registerElement is not a function
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
